fix(caja): corregir saldo esperado en cierre usando fórmula correcta
- CierreCajaViewModel: RecalcularSaldo() ahora usa la fórmula correcta:
  SaldoEsperado = MontoInicial + VentasEfectivo + IngresosEfectivo - EgresosEfectivo
  Esta fórmula incluye el Vuelto porque los movimientos de ventas (ingreso y egreso)
  ya están incluidos en los totales.

- CajaServicio: ObtenerResumenCierreAsync ahora excluye los movimientos de ventas
  (los que tienen Id_venta) de Ingresos/Egresos manuales, ya que los de ventas ya están
  incluidos en VentasEfectivo desde los pagos.

- AperturaCajaViewModel: Fix cast de MontoFinal (decimal?) a decimal.

- Agregada documentación de spec en docs/SPEC-MODULO-CAJA.md

+++ GestionComercial.UI/ViewModels/Caja/AperturaCajaViewModel.cs
@@ -146,7 +146,7 @@ namespace GestionComercial.UI.ViewModels.Caja
                 if (ultimaCajaCerrada != null)
                 {
                     UltimoCierre  = ultimaCajaCerrada.FechaCierre ?? ultimaCajaCerrada.FechaApertura;
-                    SaldoAnterior = ultimaCajaCerrada.MontoFinal; // Efectivo que quedó al cerrar
+                    SaldoAnterior = (decimal)ultimaCajaCerrada.MontoFinal; // Efectivo que quedó al cerrar
                 }
                 else
                 {
